Add aimed track detach mechanic
Manually detaching from a track is now aimed. One can get greater speed
coming off the track by aiming directly out from the track (push off the
track), or by aiming in the direction of movement on the track. The
optimal aim (for speed) is 45 degrees away from the direction moving on
the track.